// web/lib/api/models/availabilityRuleDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Reservae API
 * API para gerenciamento de espaços, disponibilidades e reservas.
 * OpenAPI spec version: v1
 */
import type { DayOfTheWeekEnum } from './dayOfTheWeekEnum';

export interface AvailabilityRuleDto {
  id?: number;
  /** @nullable */
  spaceName: string | null;
  dayOfTheWeek?: DayOfTheWeekEnum;
  startTime?: string;
  endTime?: string;
  validFrom?: string;
  validUntil?: string;
  /** @nullable */
  customPricePerSpot?: number | null;
  isActive?: boolean;
  capacity?: number;
  slotDurationMinutes?: number;
}