// web/lib/api/models/createAvailabilityRuleDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Reservae API
 * API para gerenciamento de espaços, disponibilidades e reservas.
 * OpenAPI spec version: v1
 */
import type { DayOfTheWeekEnum } from './dayOfTheWeekEnum';

export interface CreateAvailabilityRuleDTO {
  /**
     * @minimum 1
     * @maximum 2147483647
     */
  spaceId?: number;
  dayOfTheWeek?: DayOfTheWeekEnum;
  startTime?: string;
  endTime?: string;
  validFrom?: string;
  validUntil?: string;
  /**
     * @minimum 0
     * @maximum 1.7976931348623157e+308
     * @nullable
     */
  customPricePerSpot?: number | null;
  /** @nullable */
  isActive?: boolean | null;
  /**
     * @minimum 1
     * @maximum 2147483647
     */
  capacity?: number;
  /**
     * @minimum 30
     * @maximum 2147483647
     */
  slotDurationMinutes?: number;
}