// web/lib/api/models/createBookableSlotDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Reservae API
 * API para gerenciamento de espaços, disponibilidades e reservas.
 * OpenAPI spec version: v1
 */

export interface CreateBookableSlotDTO {
  /**
     * @minimum 1
     * @maximum 2147483647
     * @nullable
     */
  availabilityRuleId?: number | null;
  /**
     * @minimum 1
     * @maximum 2147483647
     * @nullable
     */
  spaceId?: number | null;
  startsAt?: string;
  endsAt?: string;
  /**
     * @minimum 0
     * @maximum 1.7976931348623157e+308
     * @nullable
     */
  customPricePerSpot?: number | null;
  /**
     * @minimum 1
     * @maximum 2147483647
     * @nullable
     */
  capacity?: number | null;
  /** @nullable */
  isActive?: boolean | null;
}